Updating Root CA MD5 by Root CA SHA256 on main CACert page
This is not a patch which I expect an approval for. This is just a means to officially ask for an update of the Root CA certificate available fro download from the main /index.php?id=3 CACert page. There should be available Root CA SHA256 signed certificate instead of the non-working obsolete Root CA MD5 certificate.