Switch database access to mysqli
This change removes all invocations of the deprecated mysql_* APIs and
replaces them with mysqli calls instead.

Configuration in includes/mysql.php is now instrumented via environment
variables. The application will not start if mandatory environment variables
are not defined. Missing environment variables are logged to the PHP/Apache
error log.

Input validation in the autocomplete backend in www/ac.php has been fixed.

The database connection is no longer stored in the session as it is only
valid for a single request and had not been reused before either.

Some TODO tags have been added at places in the code that need rework.

Add TODO.md and changelog.md as well as setup instructions in README.md.