Rename from APP to WEBDB
The environment variables containing MYSQL_APP have been renamed to
MYSQL_WEBDB to be consistent with the developer setup and to avoid
confusion with the test manager and CATS applications.

The ancient includes/.cvsignore file has been removed to avoid confusion
in IDEs.

URLs are now built consistently using the environment variables.