Implement sanitation for subject and san
This commit implements a sanity check for the SubjectDN and
SubjectAlternativeName strings before sending them to the signer.

The check uses the same regular expression as the server.pl on the
signer and will detect null bytes, illegal whitespace as well as # marks.

This is a mitigation for server.pl crashes. The failing certificate
requests are marked in the database but there is no proper feedback to the
user yet.

Detected SubjectDN and SubjectAlternative violations are logged so that a
critical admin can find the issue in the log file of the signer client.

Addresses server.pl crash from http://bugs.cacert.org/view.php?id=769